requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.1
chromadb==1.5.1
numpy<2.0.0
rank-bm25>=0.2.2
tiktoken>=0.5.0
requests>=2.31.0
huggingface_hub
typing-extensions>=4.8.0